backend/requirements.txt
gradient-adk
gradientai>=1.4.0
langchain
langchain-core
langgraph
langchain-gradient
langchain-community

fastapi[standard]
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

# PDF parsing
pdfplumber
pytesseract
Pillow
pdf2image

faiss-cpu
numpy

gradient-adk
python-dotenv
boto3 
tiktoken
langdetect
python-dotenv